RAC-392: Move register tasklet in structure
