+++ src/Akeneo/Pim/Enrichment/.php_cd.php
@@ -50,8 +50,6 @@ $rules = [
         // TODO: EASY PICK! it should be registered in the structure bundle
         'Akeneo\Pim\Structure\Bundle\DependencyInjection\Compiler\RegisterAttributeTypePass',
 
-        'Akeneo\Pim\Structure\Bundle\Manager\AttributeCodeBlacklister',
-
         // TIP-915: PIM/Enrichment should not be linked to AttributeOption
         // TIP-916: Do not check if entities exist in PQB filters or sorters
         'Akeneo\Pim\Structure\Component\Model\AttributeOptionInterface',
